requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchainhub
faiss-cpu
beautifulsoup4
duckduckgo-search
python-dotenv
unstructured
markdown
python-dotenv